debt: Refactor use-minidsp hook
- Introduced a new `useMinidsp` hook that consolidates connection, master controls, channel management, and meter handling.
