+++ jest.setup.js
@@ -1,0 +1,13 @@
+// Mock window.localStorage
+const localStorageMock = {
+  getItem: jest.fn(),
+  setItem: jest.fn(),
+  clear: jest.fn()
+};
+global.localStorage = localStorageMock;
+
+// Mock process.env
+process.env = {
+  ...process.env,
+  NEXT_PUBLIC_USE_MOCK_MINIDSP: 'true'
+};
